+++ lib/actions.js
@@ -20,10 +20,6 @@ export const getQuery = (location) => {
   return fullPath.split('?')[1] ? fullPath.split('?')[1] : ''; // return query only.
 };
 
-export const createRoute = (path, query = '') => {
-  return {path, query}
-};
-
 export const routeChange = (path) => {
   return {
     type: ROUTE_CHANGE,
