+++ lib/match.js
@@ -7,22 +7,22 @@ export const pathToRegex = (path) => {
 };
 
 const match = (path, anotherPath) => {
-  if (!path || !anotherPath) return;
+  if (!path || !anotherPath) return null;
   anotherPath = anotherPath.replace(/(\?+[|=|a-z|A-Z|0-9]*)/, '');
   const matched = anotherPath.match(pathToRegex(path));
 
   if (matched && path.match(pattern)) {
     const result = {};
     path.match(pattern).forEach((p, i) => {
-      if (!matched[i + 1]) return; // if not matched then return
+      if (!matched[i + 1]) return null; // if not matched then return
       p = p.replace(':', ''); // remove leading colon.
       result[p] = isNaN(matched[i + 1]) ? matched[i + 1] : Number(matched[i + 1]);
     });
     // return false if result is empty.
-    return Object.keys(result).length ? result : false;
+    return Object.keys(result).length ? result : null;
   }
 
-  return matched && {};
+  return matched ? {} : null;
 };
 
 export default match;
