+++ lib/router.js
@@ -11,10 +11,32 @@ import {
   getCurrent
 } from './reducer.js';
 
+import match from './match.js';
+
 let routes = {};
 
 // return cloned routes.
 export const getRoutes = () => assign({}, routes);
+
+export const findRouteKeyByPath = (path) => {
+  const routes = Object.keys(getRoutes());
+  for (let i = 0; i < routes.length; i++) {
+    const key = routes[i];
+    if (match(key, path)) return key; // return matched route.
+  }
+  return null; // return null if not found
+};
+
+export const createRoute = (path, query = '') => {
+  const route = findRouteKeyByPath(path);
+  const params = match(route, path);
+  return {
+    path,
+    route,
+    params,
+    query
+  }
+};
 
 // check and handle routeError
 export const routeErrorHandlerCreator = (store) => (data) => {
@@ -83,7 +105,7 @@ const routerCreator = (store) => {
   };
 
   const onChange = (e) => {
-    dispatch(routeChange(transformLocationToPath(location)));
+    dispatch(routeChange(createRoute(transformLocationToPath(location))));
   };
 
   // bind onChange to popstate.
